requirements.txt
Django==1.7.1
Pillow==2.6.1
South==1.0.1
argparse==1.2.1
django-widget-tweaks==1.3
wsgiref==0.1.2
django-ckeditor
